### 2018-07-31 #24 @whedon
[PRE REVIEW]: Org-Coursepack
**Submitting author:** @joonro (<a href="http://orcid.org/0000-0002-5895-163X">Joon Ro</a>)
**Repository:** <a href="https://github.com/joonro/Org-Coursepack" target ="_blank">https://github.com/joonro/Org-Coursepack</a>
**Version:** v1.0.1
