### 2018-07-31 #24 @labarba
👋  @csev  — The submitting author suggested you as editor. Since this will be your first editing assignment, let me know if you need to bounce ideas about selecting reviewers. Check out the [Editor's Guide](https://github.com/labarba/jose/blob/master/EditorsGuide.md) I put together, if you haven't. It describes the editor workflow, including needed `@whedon` commands.
