### 2018-08-10 #24 @csev
Hey @joonro - sorry for the delay.  Thanks for your submission. I like the paper and topic a lot.   I think we need to explore new ways or authoring, reusing, and remixing each other's content - and in particular doing so though github.  A couple of quick questions.  (1) There is a ?? in the paper - does that need to be resolved?  (2) looking at both the paper and your Repo, I was hoping to find a link to an example web site / course that uses this approach, and (3) if you have an example of any kind of remix / reuse of course materials I think that it would be s nice addition.  I understand that even if we prepare for (3) it does not always happen - I just thought I would ask.
