+++ paper.bib
@@ -20,7 +20,7 @@ url = {http://dx.doi.org/10.1016/j.iheduc.2013.10.001},
 volume = {20},
 year = {2014}
 }
-@book{Csikszentmihalyi_2014_Book,
+@book{Csikszentmihalyi2014Book,
 abstract = {An analysis of the reported experiences of people involved in various play-forms (i.e., rock-climbing, chess, dance, basketball, music composi- tion) suggests that the qualities which make these activities enjoyable are the following: (a) a person is able to concentrate on a limited stimulus field, (o) in which he or she can use his or her skills to meet clear demands, (c) thereby forgetting his or her own problems, and (d) his or her own separate identity, (e) at the same time obtaining a feeling of control over the environment, Q which may result in a transcendence of ego-boundaries and consequent psychic integration with metapersonal systems A formal analysis is carried out to establish what are the characteristics that an activity must have to provide such intrinsically rewarding experiences. The implications of intrinsic rewards for the understanding of human motiva- tion are briefly discussed.},
 archivePrefix = {arXiv},
 arxivId = {0803973233},
