### 2018-08-15 #24 @joonro
@csev, Regarding your question (1), I just fixed the citation with joonro/Org-Coursepack@d667599e137a044d182c7656cd5895bb098bfbb0 - I removed underscores in the BibTeX Key. I'm not sure how to update the compiled pdf file so I cannot verify that the issue is fixed at the moment, but I hope this was the issue.
