### 2018-08-17 #24 @joonro
@csev, the typo (??) in the citation is indeed fixed. Thank you for pointing this out.

Regarding your point (2), I created the documentation for Org-Coursepack as a mini-course using Org-Coursepack. The link to the course materials (https://joonro.github.io/Org-Coursepack/) is on the `Usage` section of the `README` on the repo. In addition, the compiled lecture notes are in the [Lectures directory](https://github.com/joonro/Org-Coursepack/tree/master/Courses/Org-Coursepack/2018%20Fall/Lectures). I can add links to these in the paper so they are easier to find.

In terms of (3), I have used this to create content for two courses, Marketing Research and Marketing Analytics. These two courses share a significant amount of materials, so I use and mix the modules to meet the need for each course. If these are what you are looking for, I would be happy to provide parts of them (e.g., a couple of lectures) as examples, as some course materials (e.g., exam questions) are private.

I hope I understood your questions correctly, if not, please let me know so I can provide you with further information or materials.
