### 2018-08-17 #24 @csev
@joonro - Thanks for the updates.   I just want to go from the paper to some sample course materials with the fewest possible clicks.  And while the Org-Coursepack documentation *technically* is a course - a link to a real course would be more satisfying.  I think the best solution would be to add another section to https://github.com/joonro/Org-Coursepack README.md titled "Courses that Use This" (or something similar) and have the links there.   You could put one or two links in the paper as well - but to me the most important thing is in the README.
