### 2018-08-17 #24 @joonro
@csev Understood! Thank you very much for your input - I agree I think an actual course content would be ideal here. I will let you know once I add this.
