### 2018-09-02 b7b3cd7
Add two example courses, Introduction to Regression with Excel & Python
