+++ Courses/Intro-Regression-Excel/2018 Fall/Lectures/03 Multiple Regression and Categorical Variables.tex
@@ -1,0 +1,363 @@
+% Intended LaTeX compiler: pdflatex
+\documentclass[10pt,article]{article}
+\usepackage[utf8]{inputenc}
+\usepackage[T1]{fontenc}
+\usepackage{graphicx}
+\usepackage{grffile}
+\usepackage{longtable}
+\usepackage{wrapfig}
+\usepackage{rotating}
+\usepackage[normalem]{ulem}
+\usepackage{amsmath}
+\usepackage{textcomp}
+\usepackage{amssymb}
+\usepackage{capt-of}
+\usepackage{hyperref}
+\usepackage{titling} \posttitle{\par\end{center}} \setlength{\droptitle}{-30pt} \usepackage{multicol} \setlength{\columnsep}{1cm} \usepackage[T1]{fontenc} \usepackage[utf8]{inputenc} \renewcommand{\contentsname}{Table of Contents / Agenda} \usepackage[letterpaper,left=1in,right=1in,top=0.7in,bottom=1in,headheight=23pt,includehead,includefoot,heightrounded]{geometry} \usepackage{fancyhdr} \pagestyle{fancy} \fancyhf{} \cfoot{\thepage} \usepackage{mathpazo} \usepackage[scaled=0.85]{helvet} \usepackage{courier} \usepackage[onehalfspacing]{setspace} \usepackage[framemethod=default]{mdframed} \usepackage{wrapfig} \usepackage{booktabs} \usepackage[outputdir=Lectures]{minted}
+\setcounter{secnumdepth}{3}
+\date{\vspace{-6ex}}
+\title{Class 3: Multiple Regression and Categorical Variables}
+\hypersetup{
+ pdfauthor={},
+ pdftitle={Class 3: Multiple Regression and Categorical Variables},
+ pdfkeywords={},
+ pdfsubject={Description School specific teaching materials},
+ pdfcreator={Emacs 26.1 (Org mode 9.1.13)}, 
+ pdflang={English}}
+\begin{document}
+
+\maketitle
+\lhead{ COURSE 0000 \\ Joon H. Ro } 
+\rhead{ Class 3 \\ 2018-09-04 Tue} 
+\thispagestyle{fancy}
+
+\setcounter{tocdepth}{1}
+\tableofcontents
+\vspace{6ex}
+
+\section{Categorical Variables}
+\label{sec:org462ece1}
+\subsection{Use of Dummy Variables}
+\label{sec:org05b582a}
+\begin{itemize}
+\item To capture the effect of categorical variables
+\begin{itemize}
+\item Brands, In-store displays, Gender
+\end{itemize}
+
+\item Dummy variable has a value of 0 or 1
+\begin{itemize}
+\item 1 indicates presence of characteristic
+\item 0 indicates absence of characteristic
+\end{itemize}
+\end{itemize}
+\subsection{Example}
+\label{sec:org8d318d4}
+
+{\small
+\begin{center}
+\begin{tabular}{rl}
+Sales & Store Type\\
+\hline
+10 & A\\
+4 & B\\
+8 & A\\
+6 & B\\
+7 & A\\
+6 & B\\
+7 & B\\
+8 & A\\
+\end{tabular}
+\end{center}
+}
+
+\begin{itemize}
+\item Categorical variables require recoding
+\item Use indicator variables / dummy variables
+\end{itemize}
+
+{\small
+\begin{center}
+\begin{tabular}{rlr}
+Sales & Store Type & Dummy\\
+\hline
+10 & A & 1\\
+4 & B & 0\\
+8 & A & 1\\
+6 & B & 0\\
+7 & A & 1\\
+6 & B & 0\\
+7 & B & 0\\
+8 & A & 1\\
[…]
