### 2018-09-02 #24 @joonro
Dear @csev,

Thank you again for the constructive suggestion. Per your request, I have updated the repo with two example courses: `Introduction to Regression Analysis with Excel` and `Introduction to Regression Analysis with Python.` These example courses, which are based on actual business courses that I taught in recent years, highlight the how the Org-Coursepack can help instructors reduce redundancy and increase efficiency by utilizing modularity in creating course materials.

We believe these example courses are ideal to showcase the Org-Coursepack, since they share a large portion of course materials (both courses involve teaching fundamental statistical concepts, such as hypothesis testing and regression) but also includes some course-specific materials (one course uses Excel and the other uses Python as computational tools for analyses). Users can easily see how to mix and match course content using the Org-Coursepack from these two example courses,.

Please note that we added links to these example courses in [README](https://github.com/joonro/Org-Coursepack) as you instructed.

Thank you very much for your input. Please feel free to let us know should you have any questions or requests.
