### 2018-10-23 #24 @labarba
@whedo start review
