### 2018-10-23 #34 @whedon
[REVIEW]: Org-Coursepack
**Submitting author:** @joonro (<a href="http://orcid.org/0000-0002-5895-163X">Joon Ro</a>)
**Repository:** <a href="https://github.com/joonro/Org-Coursepack" target ="_blank">https://github.com/joonro/Org-Coursepack</a>
**Version:** v1.2.0
