### 2018-10-23 #34 @labarba
@pschloss, @gregcaporaso — Thank you for agreeing to review this submission for JOSE. We're so grateful for your volunteer efforts! This is where the action happens. You will work through the checklist above, and ask questions here as needed. You may also open issues in the submission's repository, and post a link here.
