### 2018-10-29 #34 @pschloss
The manuscript describes a tool to be used in the Org mode of emacs to generate a course bundle with a syllabus, handouts, and slides. I like the idea of having a single styling that can be applied across all layers of a course packet. The authors point out that this is often done using various tools such as MS Word and PowerPoint. Another advantage is the ability to use text-based files that could be maintained under version control to track the changes to a course's content over time and foster collaboration.

To be transparent, I've never been successful in using emacs. Whether the package works as described needs to be reviewed by someone else. But, I found it difficult to navigate the documentation and templates. I could see something like this being a motivation to learn emacs and engage in this type of material in the future. Unfortunately, I find accessing the package at this level impenetrable. The installation notes state, "Clone the repo or download the ZIP archive". Again, I don't know much about emacs, but that seems pretty minimalist. Are there commands within emacs that would need to be run to load the package? Where in a directory structure in relation to the templates do I need to be to use the package? How would I build the examples? Descriptions of these steps should probably be in the README. The authors are clear that the package is for people with "basic knowledge of Org mode" and so the package is not for me. I worry that the community of people with this level of knowledge is pretty limited.

If possible, I would encourage the authors to separate the templates from the organization in their repo organization. Although the material is elsewhere, I think the README could use a more expansive description of how the repository is organized.

*Example Usage / Functionality documentation.* Could the authors perhaps create a video or set of annotated screenshots that shows how to generate one of the example course packets?

*Community guidelines.*  I would encourage the authors to create a CONTRIBUTION.md file in their repository that indicates how the community can engage with the developers to help with further development.
