### 2018-10-29 #34 @pschloss
apologies to @gregcaporaso if i accidentally checked/unchecked anything from your checklist!
