### 2018-10-30 joonro/Org-Coursepack#1 @gregcaporaso
automated tests?
I'm currently [reviewing the Org-Coursepack paper for Jose](https://github.com/openjournals/jose-reviews/issues/34).

I don't see automated tests for Org-Coursepack. Are those possible to add? It would be nice to at least see automated integration tests, where on each change, TravisCI (or some other service) runs to convert example source content into all of the different products (slides, handouts, ...).
