### 2018-10-30 joonro/Org-Coursepack#2 @gregcaporaso
contributing guidelines
I'm currently [reviewing the Org-Coursepack paper for Jose](https://github.com/openjournals/jose-reviews/issues/34).

There are no contributing guidelines that I see. At the least, it would be nice for the authors to include a CONTRIBUTING.md file.
