### 2018-10-30 joonro/Org-Coursepack#3 @gregcaporaso
quickstart guides?
I'm currently [reviewing the Org-Coursepack paper for Jose](https://github.com/openjournals/jose-reviews/issues/34).

As a non-Org-user, I find the documentation a little overwhelming. I'd like to see some sort of a quickstart guide if that's possible to develop - for example, this could cover opening one of the example courses, and then running a few commands to export slides, lecture outlines, and exams. Then, dive into the details that are covered in the [primary documentation](https://joonro.github.io/Org-Coursepack/). This would make sense to include in the _Introduction_ section, in my opinion. As it stands, I think the emacs/Org knowledge/skills required to use Org-Coursepack will be a barrier to obtaining new users.

Quickstart installation instructions could also be helpful - for example, could full installation be achieved with Anaconda, or another package manager (e.g., for the users who don't have emacs, reveal, or latex installed)? I was able to install the necessary components, but it did involve navigating to three different project websites (emacs, org-reveal, latex) to install different dependencies. This isn't essential, but it would likely help attract new users.
