### 2018-10-30 #34 @gregcaporaso
The authors present Org-Coursepack, a set of tools for deriving lecture materials (slides, handouts, syllabus, exams) using emacs/Org. I'm probably not an ideal reviewer for this work as I've never used Org, and I rarely use emacs. If the other review (@patschloss) is also not an emacs/Org user, than I think it would make sense to consider adding a third reviewer who is more familar with these tools (I avoiding reading his review while preparing mine).

I have reviewed the example lecture materials and project documentation, and the resulting course materials do seem very nicely organized. I can see how Org-Coursepack use by an instructor would be beneficial to a class.

> Version: Does the release version given match the GitHub release (v1.0.1)

The most recent release version on GitHub is 1.1.0 - that doesn't match this version number, but does match a version number.

I created a few specific issues that I think would be important to address before publication:

 - [ ] https://github.com/joonro/Org-Coursepack/issues/1

 - [ ] https://github.com/joonro/Org-Coursepack/issues/2

 - [ ] https://github.com/joonro/Org-Coursepack/issues/3
