### 2018-10-30 #34 @gregcaporaso
> apologies to @gregcaporaso if i accidentally checked/unchecked anything from your checklist!

No worries! And apologies if I checked/unchecked any of yours (I got lost in that comment block a couple of times and worried that I may have messed with your list).
