### 2018-11-19 joonro/Org-Coursepack#1 @joonro
Hi -

I'm sorry for the delay. I looked into incorporating TravisCI, but given that `org-coursepack` is a template mainly consists of files written in org-markup, it was not clear which part I should make tests for, except for the `Update Lectures` script which automatically updates class number, class agenda, etc for each class.

Could you please advise us the kind of tests that we can provide? Please let me know and thank you very much.
