### 2018-11-19 ca3b598
Add Contributing Guidelines
