+++ CONTRIBUTING.md
@@ -1,0 +1,38 @@
+
+# Contributing to Org-Coursepack
+
+Org-Coursepack is an open source project and we appreciate community
+contributions. Contributions should be made in the form of GitHub
+pull requests.
+
+
+## How Org-Coursepack is written
+
+Org-Coursepack consists of multiple Org files written in
+[org-mode](<https://orgmode.org/>), a major mode in Emacs. Hence, when you're
+submitting changes to content, you'll be submitting changes to Org files.
+
+In general we use [Git-flow](https://www.atlassian.com/git/tutorials/comparing-workflows/gitflow-workflow) workflow, so we use `master` for releases. Hence,
+[pull requests](https://help.github.com/articles/about-pull-requests/) should be made against the `develop` branch.
+
+
+## Small Contributions
+
+Since org-mode uses a markup syntax similar to markdown and reStructuredText,
+small changes such as fixing typos and changing wording can be easily made
+directly on GitHub using its online editor, by clicking on `Edit this file`
+link while viewing an individual file. You can create a pull request from the
+page once editing is done.
+
+
+## Large Contributions
+
+If you are interested in making substantial changes on Org-Coursepack,
+please first open an issue so the issue can discussed before the changes are
+made.
+
+
+## All set!
+
+If you have any questions feel free to post them to <https://github.com/joonro/Org-Coursepack/issues>.
+
