### 2018-11-19 51c28d2
Made changes to the flow and content
