+++ CONTRIBUTING.md
@@ -1,38 +1,33 @@
 
-# Contributing to Org-Coursepack
+# Contributing to the Org-Coursepack
 
-Org-Coursepack is an open source project and we appreciate community
+Org-Coursepack is an open source project, and we appreciate community
 contributions. Contributions should be made in the form of GitHub
 pull requests.
 
-
-## How Org-Coursepack is written
-
-Org-Coursepack consists of multiple Org files written in
-[org-mode](<https://orgmode.org/>), a major mode in Emacs. Hence, when you're
-submitting changes to content, you'll be submitting changes to Org files.
-
-In general we use [Git-flow](https://www.atlassian.com/git/tutorials/comparing-workflows/gitflow-workflow) workflow, so we use `master` for releases. Hence,
-[pull requests](https://help.github.com/articles/about-pull-requests/) should be made against the `develop` branch.
-
+In general, we follow [Git-flow](https://www.atlassian.com/git/tutorials/comparing-workflows/gitflow-workflow) 
+as our workflow; hence, the `master` branch is used for releases, 
+and [pull requests](https://help.github.com/articles/about-pull-requests/) 
+should be based on the `develop` branch.
 
 ## Small Contributions
 
-Since org-mode uses a markup syntax similar to markdown and reStructuredText,
-small changes such as fixing typos and changing wording can be easily made
-directly on GitHub using its online editor, by clicking on `Edit this file`
+Org-Coursepack consists of multiple Org files written in
+[Org-mode](<https://orgmode.org/>), a major mode in Emacs. Since Org-mode uses 
+a markup syntax similar to Markdown and reStructuredText,
+small changes (e.g., fixing typos and wording) can be made
+directly on GitHub using its online editor by clicking the `Edit this file`
 link while viewing an individual file. You can create a pull request from the
-page once editing is done.
+page once you finish editing.
 
 
 ## Large Contributions
 
-If you are interested in making substantial changes on Org-Coursepack,
-please first open an issue so the issue can discussed before the changes are
-made.
+If you are interested in making substantial changes to the Org-Coursepack,
+please first open an issue so the issue can be discussed prior to making changes.
 
 
 ## All set!
 
-If you have any questions feel free to post them to <https://github.com/joonro/Org-Coursepack/issues>.
+If you have any questions, feel free to post them to <https://github.com/joonro/Org-Coursepack/issues>.
 
