### 2018-11-19 joonro/Org-Coursepack#2 @joonro
Dear @gregcaporaso, we just added [CONTRIBUTING.md](https://github.com/joonro/Org-Coursepack/blob/feature/CONTRIBUTING.md/CONTRIBUTING.md) (commit: 51c28d283995fdac8ce6f54d9284c03751b6f5ad). Please let me know if you need anything else regarding this issue. We will work on making the quickstarter guide. Thank you very much!
