### 2018-11-19 #34 @joonro
Dear @pschloss, we just added [CONTRIBUTING.md](https://github.com/joonro/Org-Coursepack/blob/feature/CONTRIBUTING.md/CONTRIBUTING.md) (commit: https://github.com/joonro/Org-Coursepack/commit/51c28d283995fdac8ce6f54d9284c03751b6f5ad) to address your comment regarding _Community guidelines_. Please let me know if you need anything else regarding this issue.

We will add the quickstarter guide with screenshots. We believe that will address your comment regarding _Example Usage / Functionality documentation_. Thank you very much!
