### 2018-11-27 #34 @pschloss
Thanks for adding the CONTRIBUTING.md file. I'm not seeing the quickstarter guide. Given that neither @gregcaporaso nor I have much experience with OrgMode or emacs, it would be ideal if @labarba could find someone to finish the review who does have experience. Perhaps Naupaka Zimmerman (gh: naupaka) or Kara Woo (gh: karawoo) would be up for it?
