### 2018-11-27 #34 @joonro
Yes we are working on the quickstarter guide. I will report back when it is done!
