### 2018-12-03 374b67c
First draft of the quickstarter guide
