+++ Courses/Org-Coursepack-Quickstart-Guide/2018 Fall/Lectures/01 Introduction.tex
@@ -1,0 +1,99 @@
+% Intended LaTeX compiler: pdflatex
+\documentclass[10pt,article]{article}
+\usepackage[utf8]{inputenc}
+\usepackage[T1]{fontenc}
+\usepackage{graphicx}
+\usepackage{grffile}
+\usepackage{longtable}
+\usepackage{wrapfig}
+\usepackage{rotating}
+\usepackage[normalem]{ulem}
+\usepackage{amsmath}
+\usepackage{textcomp}
+\usepackage{amssymb}
+\usepackage{capt-of}
+\usepackage{hyperref}
+\usepackage{titling} \posttitle{\par\end{center}} \setlength{\droptitle}{-30pt} \usepackage{multicol} \setlength{\columnsep}{1cm} \usepackage[T1]{fontenc} \usepackage[utf8]{inputenc} \renewcommand{\contentsname}{Table of Contents / Agenda} \usepackage[letterpaper,left=1in,right=1in,top=0.7in,bottom=1in,headheight=23pt,includehead,includefoot,heightrounded]{geometry} \usepackage{fancyhdr} \pagestyle{fancy} \fancyhf{} \cfoot{\thepage} \usepackage{mathpazo} \usepackage[scaled=0.85]{helvet} \usepackage{courier} \usepackage[onehalfspacing]{setspace} \usepackage[framemethod=default]{mdframed} \usepackage{wrapfig} \usepackage{booktabs} \usepackage[outputdir=Lectures]{minted}
+\setcounter{secnumdepth}{3}
+\date{\vspace{-6ex}}
+\title{Class 1: Introduction}
+\hypersetup{
+ pdfauthor={},
+ pdftitle={Class 1: Introduction},
+ pdfkeywords={},
+ pdfsubject={Org-Coursepack quickstart guide School specific teaching materials},
+ pdfcreator={Emacs 26.1 (Org mode 9.1.14)}, 
+ pdflang={English}}
+\begin{document}
+
+\maketitle
+\lhead{ COURSE 0000 \\ Your Name } 
+\rhead{ Class 1 \\ 2018-08-28 Tue} 
+\thispagestyle{fancy}
+
+\setcounter{tocdepth}{1}
+\tableofcontents
+\vspace{6ex}
+
+\section{Introduction to Org-Coursepack Quickstart guide}
+\label{sec:org4a2b6c9}
+Content for this section here.
+\section{New section}
+\label{sec:orgc9353ff}
+\subsection{Lists}
+\label{sec:org9ba10fb}
+Obviously you cannot use \texttt{*} to specify a list, but otherwise Org mode
+uses a typical syntax for lists. For example,
+
+\begin{itemize}
+\item List
+\begin{itemize}
+\item Nested list item 1
+\item Nested list item 2
+\end{itemize}
+\end{itemize}
+
+
+\begin{enumerate}
+\item Numbered list
+\end{enumerate}
+
+\subsection{Math}
+\label{sec:org88a9dc3}
+You can directly input \LaTeX{} math in Org mode. For example,
+
+\[ \cos (2\theta) = \cos^2 \theta - \sin^2 \theta \]
+\subsection{Slide split}
+\label{sec:org6a4cbce}
+Users can put \texttt{\#+REVEAL: split} to split a slide. For example,
+
+This line will be shown in a new slide.
+\subsection{Fragmented Contents}
+\label{sec:org0a59814}
+\begin{itemize}
+\item This list
+\item is fragmented
+\item in reveal.js slides
+\end{itemize}
+\subsection{Local Images}
+\label{sec:org42ec809}
+Local images can be included in this way:
+
+\begin{center}
+\includegraphics[width=4cm]{../../../Assets/Images/Misc/affiche.png}
+\end{center}
+
[…]
