+++ Courses/Org-Coursepack-Quickstart-Guide/2018 Fall/Exams/Exam_1_01_A.tex
@@ -1,0 +1,78 @@
+% Intended LaTeX compiler: pdflatex
+\documentclass[article,letterpaper,times,10pt,listings-bw,microtype]{article}
+\usepackage[utf8]{inputenc}
+\usepackage[T1]{fontenc}
+\usepackage{graphicx}
+\usepackage{grffile}
+\usepackage{longtable}
+\usepackage{wrapfig}
+\usepackage{rotating}
+\usepackage[normalem]{ulem}
+\usepackage{amsmath}
+\usepackage{textcomp}
+\usepackage{amssymb}
+\usepackage{capt-of}
+\usepackage{hyperref}
+\usepackage{titling} \usepackage{multicol} \setlength{\columnsep}{1cm} \usepackage[T1]{fontenc} \usepackage{mathpazo} \usepackage[scaled=0.85]{helvet} \usepackage{courier} \usepackage[singlespacing]{setspace} \usepackage{geometry}\geometry{left=1in,right=1in,top=0.7in,bottom=1in} \usepackage[framemethod=default]{mdframed} \usepackage{enumitem} \setlist[1]{itemsep=15pt}
+\author{COURSE 0000-01: Template}
+\date{2018-02-27 Tue}
+\title{Exam 1 (Type A)}
+\hypersetup{
+ pdfauthor={COURSE 0000-01: Template},
+ pdftitle={Exam 1 (Type A)},
+ pdfkeywords={},
+ pdfsubject={Org-Coursepack-Documentation: A Comprehensive Course Template for Teaching with Org Mode School specific teaching materials},
+ pdfcreator={Emacs 25.3.1 (Org mode 9.1.13)}, 
+ pdflang={English}}
+\begin{document}
+
+\maketitle
+\vspace{20 mm}
+
+\begin{itemize}
+\item Your Name: \underline{\hspace{4cm}}
+\end{itemize}
+
+\vspace{20 mm}
+
+Instructions:
+
+\begin{itemize}
+\item Total time: 75 Minutes
+\item There are both multiple choice and short answer questions
+\end{itemize}
+\clearpage
+\section*{Multiple Choice}
+\label{sec:org7d6dbda}
+(2 points each, unless otherwise specified)
+\begin{enumerate}
+\item Multiple Choice Question.
+\begin{enumerate}
+\item Choice
+\item Choice
+\item Choice
+\end{enumerate}
+
+\item Multiple Choice Question.
+\begin{enumerate}
+\item Choice
+\item Choice
+\item Choice
+\end{enumerate}
+\end{enumerate}
+\clearpage
+\section*{Short Answers}
+\label{sec:org1755fad}
+\begin{mdframed}[style=exampledefault, frametitle={}]
+Please answer the following short answer questions each with no more than 3
+sentences.
+\end{mdframed}
+
+\begin{enumerate}
+\item Describe how to include the content of a subtree from another org file, excluding the heading
+
+\vspace{50 mm}
+
+\item Describe how to include the content of a subtree from another org file, including the heading
+\end{enumerate}
+\end{document}

[…]
