+++ paper.bib
@@ -1,37 +1,0 @@
-Automatically generated by Mendeley Desktop 1.19.1
-Any changes to this file will be lost if it is regenerated by Mendeley.
-
-BibTeX export options can be customized via Options -> BibTeX in Mendeley Desktop
-
-@article{Kuo2014,
-abstract = {Student satisfaction is important in the evaluation of distance education courses as it is related to the quality of online programs and student performance. Interaction is a critical indicator of student satisfaction; however, its impact has not been tested in the context of other critical student- and class-level predictors. In this study, we tested a regression model for student satisfaction involving student characteristics (three types of interaction, Internet self-efficacy, and self-regulated learning) and class-level predictors (course category and academic program). Data were collected in a sample of 221 graduate and undergraduate students responding to an online survey. The regression model was tested using hierarchical linear modeling (HLM). Learner-instructor interaction and learner-content interaction were significant predictors of student satisfaction but learner-learner interaction was not. Learner-content interaction was the strongest predictor. Academic program category moderated the effect of learner-content interaction on student satisfaction. The effect of learner-content interaction on student satisfaction was stronger in Instructional Technology and Learning Sciences than in psychology, physical education or family, consumer, and human development. In sum, the results suggest that improvements in learner-content interaction yield most promise in enhancing student satisfaction and that learner-learner interaction may be negligible in online course settings.},
-author = {Kuo, Yu Chun and Walker, Andrew E. and Schroder, Kerstin E.E. and Belland, Brian R.},
-doi = {10.1016/j.iheduc.2013.10.001},
-file = {:home/joon/Documents/Mendeley Library/Internet and Higher Education/2014/Kuo et al. - 2014 - Internet and Higher Education.pdf:pdf},
-isbn = {1096-7516},
-issn = {10967516},
-journal = {Internet and Higher Education},
-keywords = {Hierarchical linear modeling (HLM),Interactions,Internet self-efficacy,Online education,Satisfaction,Self-regulated learning},
-pages = {35--50},
-pmid = {24921255},
-publisher = {Elsevier B.V.},
-title = {{Interaction, Internet self-efficacy, and self-regulated learning as predictors of student satisfaction in online education courses}},
-url = {http://dx.doi.org/10.1016/j.iheduc.2013.10.001},
-volume = {20},
-year = {2014}
-}
-@book{Csikszentmihalyi2014Book,
-abstract = {An analysis of the reported experiences of people involved in various play-forms (i.e., rock-climbing, chess, dance, basketball, music composi- tion) suggests that the qualities which make these activities enjoyable are the following: (a) a person is able to concentrate on a limited stimulus field, (o) in which he or she can use his or her skills to meet clear demands, (c) thereby forgetting his or her own problems, and (d) his or her own separate identity, (e) at the same time obtaining a feeling of control over the environment, Q which may result in a transcendence of ego-boundaries and consequent psychic integration with metapersonal systems A formal analysis is carried out to establish what are the characteristics that an activity must have to provide such intrinsically rewarding experiences. The implications of intrinsic rewards for the understanding of human motiva- tion are briefly discussed.},
-archivePrefix = {arXiv},
-arxivId = {0803973233},
-author = {Csikszentmihalyi, Mihaly},
-doi = {10.1007/978-94-017-9088-8},
-eprint = {0803973233},
-file = {:home/joon/Documents/Mendeley Library/Unknown/2014/Csikszentmihalyi - 2014 - Unknown.pdf:pdf},
-isbn = {978-94-017-9087-1},
-issn = {0022-1678},
-pmid = {4852177},
-title = {{Flow And The Foundations Of Positive Psychology}},
-url = {http://link.springer.com/10.1007/978-94-017-9088-8},
-year = {2014}
-}
