+++ sphinx/Lectures/09 Semester Org Files 4 4 Exams and Assignments.rst
@@ -1,2 +1,0 @@
-.. include:: ../../Courses/Org-Coursepack/2018 Fall/Lectures/09 Semester Org Files 4 4 Exams and Assignments.rst
-

[…]
