+++ _sources/index.rst.txt
@@ -1,0 +1,29 @@
+.. Org-Coursepack documentation master file, created by
+   sphinx-quickstart on Tue Jul 24 09:43:16 2018.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+Welcome to Org-Coursepack's documentation!
+==========================================
+
+.. toctree::
+   :maxdepth: 2
+   :caption: Contents:
+
+   Lectures/01 Introduction.rst
+   Lectures/02 Org Mode Basics.rst
+   Lectures/03 Topic and Course Org Files.rst
+   Lectures/04 Creating Content for Slides and Handouts.rst
+   Lectures/05 Exporting Slides and Handouts.rst
+   Lectures/06 Semester Org Files 1 4 Introduction.rst
+   Lectures/07 Semester Org Files 2 4 Sections and Syllabus.rst
+   Lectures/08 Semester Org Files 3 4 Lectures.rst
+   Lectures/09 Semester Org Files 4 4 Exams and Assignments.rst
+
+
+Indices and tables
+==================
+
+* :ref:`genindex`
+* :ref:`modindex`
+* :ref:`search`
