+++ docs/_sources/Lectures/06 Semester Org Files 1 4 Introduction.rst.txt
@@ -1,2 +1,0 @@
-.. include:: ../../Courses/Org-Coursepack/2018 Fall/Lectures/06 Semester Org Files 1 4 Introduction.rst
-
