+++ Courses/Org-Coursepack/2018 Fall/Lectures/01 Introduction.rst
@@ -59,11 +59,414 @@ materials using `Org mode <https://orgmode.org/manual/Export-settings.html>`_, a
 1.3 Requirements
 ~~~~~~~~~~~~~~~~
 
-- Basic knowledge of Emacs
+- Org-Coursepack is a tool based on `Org mode <https://orgmode.org/manual/Export-settings.html>`_, a major mode in `GNU Emacs <https://www.gnu.org/software/emacs/manual/html_node/emacs/Specifying-File-Variables.html#Specifying-File-Variables>`_. It
+  does not, however, require extensive previous experience with either, unless
+  the user wants to make changes to pre-built scripts. The basic level of
+  knowledge needed is:
 
-- Basic knowledge of Org mode
+  Emacs
+      Basic knowledge of text editing using Emacs
 
-2 Overview of the Directory Structure
+  Org mode
+      Basic knowledge of Org mode markup syntax. Org mode uses a
+      straightforward markup language similar to Markdown and
+      reStructuredText, thus it will be easy to learn for any user
+      who is familiar with other markup languages.
+
+2 Quickstart Guide
+------------------
+
+In this quickstart guide, we will guide you through how to create your course
+with ``Org-Coursepack`` with step-by-step examples. We focus on syllabi and lectures, 
+as they tend to make up for the majority of the course content, and other parts such as 
+assignments and exams can be created similarly.
+
+2.1 Copy and rename ``Template`` course
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+We use ``/`` as the root path of Org-Coursepack throughout this quickstart
+guide. Make a copy of the ``Template`` directory inside ``/Courses`` folder and
+rename the folder and ``template.org`` inside the folder to your course name. We
+call this ``Org-Coursepack-Quickstart-Guide`` in this tutorial:
+
+.. image:: ../../../Assets/Images/Org-Teaching/Quickstart_Course-renamed.png
+
+Your specific course for a semester will reside inside ``Semester`` folder. We
+assume that you are preparing a course for the fall 2018 semester -- let's
+rename the folder and ``Semestor.org`` inside the folder to ``2018 Fall`` and
+``2018 Fall.org``, respectively. From this semester Org file you will be
+exporting your actual course content.
+
+.. image:: ../../../Assets/Images/Org-Teaching/Quickstart_Semester-renamed.png
+
+2.2 Set local variables permission in semester Org file
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+Visit (i.e., open) ``2018 Fall.org``. First time you visit this file, it will 
+show you the following warning about local variables:
+
+.. code:: common-lisp
+
+    The local variables list in 201 Fall.org
+    contains values that may not be safe (*), and variables that are risky (**)
+
+These variables have time-format settings, as well as
+``org-confirm-elisp-link-function: nil``, which allows you to click on links
+(mainly for exporting) without being asked to confirm. You can type ``!`` to set
+them permanently for your convenience.
+
+Then you will see that the file is open and it has a template to construct a
+course for this semester:
+
+.. image:: ../../../Assets/Images/Org-Teaching/Quickstart_Semester-Org-File.png
+
+You can freely move around with movement keys. Drawers (e.g., ``:LOCALSETUP:``)
+and subtrees (e.g., ``\* Sections``) under the cursor can be expanded and
+collapsed by pressing ``Tab`` key:
+
+.. image:: ../../../Assets/Images/Org-Teaching/Quickstart_Org-Expand.png
+
[…]
