+++ docs/_sources/Lectures/01 Introduction.rst.txt
@@ -1,2 +1,0 @@
-.. include:: ../../Courses/Org-Coursepack/2018 Fall/Lectures/01 Introduction.rst
-
