+++ docs/_sources/Lectures/05 Exporting Slides and Handouts.rst.txt
@@ -1,2 +1,0 @@
-.. include:: ../../Courses/Org-Coursepack/2018 Fall/Lectures/05 Exporting Slides and Handouts.rst
-
