+++ docs/_sources/Syllabus/Syllabus (Section 1).rst.txt
@@ -1,1 +1,0 @@
-.. include:: ../../Courses/Org-Coursepack/2018 Fall/Syllabus/Syllabus (Section 1).rst
