+++ Courses/Org-Coursepack-Quickstart-Guide/2018 Fall/Lectures/01 Introduction.tex
@@ -36,12 +36,12 @@
 \vspace{6ex}
 
 \section{Introduction to Org-Coursepack Quickstart guide}
-\label{sec:org4a2b6c9}
+\label{sec:orgf7bfc3b}
 Content for this section here.
 \section{New section}
-\label{sec:orgc9353ff}
+\label{sec:org343d404}
 \subsection{Lists}
-\label{sec:org9ba10fb}
+\label{sec:org2b934c8}
 Obviously you cannot use \texttt{*} to specify a list, but otherwise Org mode
 uses a typical syntax for lists. For example,
 
@@ -59,24 +59,24 @@ uses a typical syntax for lists. For example,
 \end{enumerate}
 
 \subsection{Math}
-\label{sec:org88a9dc3}
+\label{sec:org00c21b2}
 You can directly input \LaTeX{} math in Org mode. For example,
 
 \[ \cos (2\theta) = \cos^2 \theta - \sin^2 \theta \]
 \subsection{Slide split}
-\label{sec:org6a4cbce}
+\label{sec:org8c23102}
 Users can put \texttt{\#+REVEAL: split} to split a slide. For example,
 
 This line will be shown in a new slide.
 \subsection{Fragmented Contents}
-\label{sec:org0a59814}
+\label{sec:org6b0e999}
 \begin{itemize}
 \item This list
 \item is fragmented
 \item in reveal.js slides
 \end{itemize}
 \subsection{Local Images}
-\label{sec:org42ec809}
+\label{sec:orgc3748cd}
 Local images can be included in this way:
 
 \begin{center}
@@ -84,7 +84,7 @@ Local images can be included in this way:
 \end{center}
 
 \subsection{Web Images / Hiding Contents}
-\label{sec:orgd6f9897}
+\label{sec:orgd43f256}
 Remote images can be included by directly using its URL. 
 
 \iffalse
