+++ Courses/Org-Coursepack/2018 Fall/Lectures/01 Course Introduction.rst
@@ -1,0 +1,582 @@
+============================
+Class 1: Course Introduction
+============================
+
+
+ ORG 0000  Joon H. Ro & Jae-Eun Namkoong  
+ Class 1  [2018-08-28 Tue] 
+
+1 Introduction to Org-Coursepack
+--------------------------------
+
+The Org-Coursepack provides a template for developing and managing teaching
+materials using `Org mode <https://orgmode.org/manual/Export-settings.html>`_, a major mode in `GNU Emacs <https://www.gnu.org/software/emacs/manual/html_node/emacs/Specifying-File-Variables.html#Specifying-File-Variables>`_.
+
+1.1 Advantages for Instructors
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+- First, Org mode and modular design allow for more effective and efficient
+  content creation.
+
+  - Content updates get propagated across courses, semesters, and sections,
+    minimizing the potential for inconsistencies
+
+  - Minimizes redundancy when sharing content across courses, semesters, and
+    sections
+
+- Second, instructors can enjoy the benefits of having a flexible export system and an output-specific export option.
+
+  - Consistent content across multiple output formats
+
+    - Slides (e.g., via reveal.js or Beamer backends)
+
+    - Handouts (e.g., via the LaTeX or reStructuredText backends)
+
+  - Selective formatting and presentation of components depending on output
+    format
+
+- Third, the template contains a) utility functions written in Emacs Lisp, b)
+  shortcuts to Org mode functions, and c) pre-built tree structures, which
+  allow automation of many tasks including:
+
+  - Automatic class numbering
+
+  - Automatic creation of key content including (but not limited to)
+
+    - course schedule for syllabi;
+
+    - agenda of lecture materials; and
+
+    - exam keys.
+
+1.2 Advantages for Students
+~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+- Consistent, properly-formatted, and strategically presented course materials add to student engagement
+
+- Availability of materials that are easier to digest and review outside the classroom
+
+1.3 Requirements
+~~~~~~~~~~~~~~~~
+
+- Org-Coursepack is a tool based on `Org mode <https://orgmode.org/manual/Export-settings.html>`_, a major mode in `GNU Emacs <https://www.gnu.org/software/emacs/manual/html_node/emacs/Specifying-File-Variables.html#Specifying-File-Variables>`_. It
+  does not, however, require extensive previous experience with either, unless
+  the user wants to make changes to pre-built scripts. Only the following basic knowledge is needed:
+
+  Emacs
+      Text editing using Emacs
+
+  Org mode
+      Org mode markup syntax. Org mode uses a
+      straightforward markup language similar to Markdown and
+      reStructuredText, and thus, it will be easy to learn for any user
+      who is familiar with other markup languages.
+
+2 Quickstart Guide
+------------------
+
+This quickstart guide will show you, step-by-step, how to create your course
+with ``Org-Coursepack``. We focus on creating syllabi and lectures, 
+as they typically make up the majority of course content. Other course materials, such as 
+assignments and exams, can be created in similar ways.
+
+2.1 Installation
+~~~~~~~~~~~~~~~~
+
+Follow instructions on `README <https://github.com/joonro/Org-Coursepack#installation>`_ to install ``Org-Coursepack`` along with its
+dependencies.
+
+2.2 Copy and rename the ``Template`` directory and Org file
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+The quickstart guide uses ``/`` to refer the root path of ``Org-Coursepack``. Make
[…]
