+++ Courses/Org-Coursepack/2018 Fall/Lectures/01 Introduction.tex
@@ -1,673 +1,0 @@
-% Intended LaTeX compiler: pdflatex
-\documentclass[10pt,article]{article}
-\usepackage[utf8]{inputenc}
-\usepackage[T1]{fontenc}
-\usepackage{graphicx}
-\usepackage{grffile}
-\usepackage{longtable}
-\usepackage{wrapfig}
-\usepackage{rotating}
-\usepackage[normalem]{ulem}
-\usepackage{amsmath}
-\usepackage{textcomp}
-\usepackage{amssymb}
-\usepackage{capt-of}
-\usepackage{hyperref}
-\usepackage{titling} \posttitle{\par\end{center}} \setlength{\droptitle}{-30pt} \usepackage{multicol} \setlength{\columnsep}{1cm} \usepackage[T1]{fontenc} \usepackage[utf8]{inputenc} \renewcommand{\contentsname}{Table of Contents / Agenda} \usepackage[letterpaper,left=1in,right=1in,top=0.7in,bottom=1in,headheight=23pt,includehead,includefoot,heightrounded]{geometry} \usepackage{fancyhdr} \pagestyle{fancy} \fancyhf{} \cfoot{\thepage} \usepackage{mathpazo} \usepackage[scaled=0.85]{helvet} \usepackage{courier} \usepackage[onehalfspacing]{setspace} \usepackage[framemethod=default]{mdframed} \usepackage{wrapfig} \usepackage{booktabs} \usepackage[outputdir=Lectures]{minted}
-\setcounter{secnumdepth}{3}
-\date{\vspace{-6ex}}
-\title{Class 1: Introduction}
-\hypersetup{
- pdfauthor={},
- pdftitle={Class 1: Introduction},
- pdfkeywords={},
- pdfsubject={Fall 2018 Org-Coursepack School specific teaching materials},
- pdfcreator={Emacs 26.1 (Org mode 9.1.14)}, 
- pdflang={English}}
-\begin{document}
-
-\maketitle
-\lhead{ ORG 0000 \\ Joon H. Ro \& Jae-Eun Namkoong } 
-\rhead{ Class 1 \\ 2018-08-28 Tue} 
-\thispagestyle{fancy}
-
-\setcounter{tocdepth}{1}
-\tableofcontents
-\vspace{6ex}
-
-\section{Introduction to Org-Coursepack}
-\label{sec:org9bff693}
-The Org-Coursepack provides a template for developing and managing teaching
-materials using \href{https://orgmode.org/manual/Export-settings.html}{Org mode}, a major mode in \href{https://www.gnu.org/software/emacs/manual/html\_node/emacs/Specifying-File-Variables.html\#Specifying-File-Variables}{GNU Emacs}.
-\subsection{Advantages for Instructors}
-\label{sec:orgbf83241}
-\begin{itemize}
-\item First, Org mode and modular design allow for more effective and efficient
-content creation.
-\begin{itemize}
-\item Content updates get propagated across courses, semesters, and sections,
-minimizing the potential for inconsistencies
-\item Minimizes redundancy when sharing content across courses, semesters, and
-sections
-\end{itemize}
-\end{itemize}
-
-\begin{itemize}
-\item Second, instructors can enjoy the benefits of having a flexible export system and an output-specific export option.
-\begin{itemize}
-\item Consistent content across multiple output formats
-\begin{itemize}
-\item Slides (e.g., via reveal.js or Beamer backends)
-\item Handouts (e.g., via the \LaTeX{} or reStructuredText backends)
-\end{itemize}
-\item Selective formatting and presentation of components depending on output
-format
-\end{itemize}
-\item Third, the template contains a) utility functions written in Emacs Lisp, b)
-shortcuts to Org mode functions, and c) pre-built tree structures, which
[…]
