+++ _sources/Lectures/01 Course Introduction.rst.txt
@@ -1,0 +1,2 @@
+.. include:: ../../Courses/Org-Coursepack/2018 Fall/Lectures/01 Course Introduction.rst
+
