### 2018-12-25 98659f9
Stop tacking .tex files
