+++ Courses/Org-Coursepack/2018 Fall/Lectures/08 Semester Org Files 3 4 Lectures.tex
@@ -1,292 +1,0 @@
-% Intended LaTeX compiler: pdflatex
-\documentclass[10pt,article]{article}
-\usepackage[utf8]{inputenc}
-\usepackage[T1]{fontenc}
-\usepackage{graphicx}
-\usepackage{grffile}
-\usepackage{longtable}
-\usepackage{wrapfig}
-\usepackage{rotating}
-\usepackage[normalem]{ulem}
-\usepackage{amsmath}
-\usepackage{textcomp}
-\usepackage{amssymb}
-\usepackage{capt-of}
-\usepackage{hyperref}
-\usepackage{titling} \posttitle{\par\end{center}} \setlength{\droptitle}{-30pt} \usepackage{multicol} \setlength{\columnsep}{1cm} \usepackage[T1]{fontenc} \usepackage[utf8]{inputenc} \renewcommand{\contentsname}{Table of Contents / Agenda} \usepackage[letterpaper,left=1in,right=1in,top=0.7in,bottom=1in,headheight=23pt,includehead,includefoot,heightrounded]{geometry} \usepackage{fancyhdr} \pagestyle{fancy} \fancyhf{} \cfoot{\thepage} \usepackage{mathpazo} \usepackage[scaled=0.85]{helvet} \usepackage{courier} \usepackage[onehalfspacing]{setspace} \usepackage[framemethod=default]{mdframed} \usepackage{wrapfig} \usepackage{booktabs} \usepackage[outputdir=Classes]{minted}
-\setcounter{secnumdepth}{3}
-\date{\vspace{-6ex}}
-\title{Class 8: Semester Org Files (3/4): Lectures}
-\hypersetup{
- pdfauthor={},
- pdftitle={Class 8: Semester Org Files (3/4): Lectures},
- pdfkeywords={},
- pdfsubject={Org-Coursepack-Documentation: A Comprehensive Course Template for Teaching with Org Mode School specific teaching materials},
- pdfcreator={Emacs 25.3.1 (Org mode 9.1.13)}, 
- pdflang={English}}
-\begin{document}
-
-\maketitle
-\lhead{ ORG 0000 \\ Joon H. Ro \& Jae-Eun Namkoong } 
-\rhead{ Class 8 \\ 2018-09-20 Thu} 
-\thispagestyle{fancy}
-
-\setcounter{tocdepth}{1}
-\tableofcontents
-\vspace{6ex}
-
-\section{The \texttt{Lectures} Top-level Tree}
-\label{sec:orgdb3c315}
-\begin{itemize}
-\item The \texttt{Lectures} top-level tree, which is the basis for the class
-schedule table, contains a subtree for each class and event (e.g.,
-assignment, exam). See the example below.
-
-{\small
-\begin{verbatim}
-  * Lectures
-  :PROPERTIES:
-  
-  #+NAME: Update Classes
-  #+BEGIN_SRC emacs-lisp :results none
-  
-  ** Tasks [0/1]                                           :noexport:skipcount:
-  ** Classes and Assignments Dates                                  :skipcount:
-  ** Introduction
-  ** Org Mode Basics
-  ** ...
-  ** _Assignment 1 Due_                     :Assignment:skipcount:
-  ** _Exam 1_                                                           :Exam:
-  ** *Thanksgiving Holiday*                                 :Holiday:skipcount:
-  ** _Final Exam_                                                   :skipcount:
-\end{verbatim}
-}
-\end{itemize}
-\subsection{\texttt{:PROPERTIES:}}
-\label{sec:orgb2921a9}
-\begin{itemize}
-\item The properties of the \texttt{Lectures} top-level tree contain information common
-across lectures such as export options, including the handout \LaTeX{}
[…]
