### 2018-12-25 5964efd
Rename the first lecture
