### 2018-12-25 e2eaf42
Remove obsolete course quickstart guide
