+++ Courses/Org-Coursepack-Quickstart-Guide/2018 Fall/Exams/Exam_1_MC_A_Key.tex
@@ -1,45 +1,0 @@
-% Intended LaTeX compiler: pdflatex
-\documentclass[article,letterpaper,times,10pt,listings-bw,microtype]{article}
-\usepackage[utf8]{inputenc}
-\usepackage[T1]{fontenc}
-\usepackage{graphicx}
-\usepackage{grffile}
-\usepackage{longtable}
-\usepackage{wrapfig}
-\usepackage{rotating}
-\usepackage[normalem]{ulem}
-\usepackage{amsmath}
-\usepackage{textcomp}
-\usepackage{amssymb}
-\usepackage{capt-of}
-\usepackage{hyperref}
-\usepackage{titling} \usepackage{multicol} \setlength{\columnsep}{1cm} \usepackage[T1]{fontenc} \usepackage{mathpazo} \usepackage[scaled=0.85]{helvet} \usepackage{courier} \usepackage[singlespacing]{setspace} \usepackage{geometry}\geometry{left=1in,right=1in,top=0.7in,bottom=1in} \usepackage[framemethod=default]{mdframed} \usepackage{enumitem} \setlist[1]{itemsep=15pt}
-\author{COURSE 0000: Template \\ Semester}
-\date{2018-02-27 Tue}
-\title{Exam 1 (Type A) Key}
-\hypersetup{
- pdfauthor={COURSE 0000: Template \\ Semester},
- pdftitle={Exam 1 (Type A) Key},
- pdfkeywords={},
- pdfsubject={Org-Coursepack-Documentation: A Comprehensive Course Template for Teaching with Org Mode School specific teaching materials},
- pdfcreator={Emacs 25.3.1 (Org mode 9.1.13)}, 
- pdflang={English}}
-\begin{document}
-
-\maketitle
-\begin{enumerate}
-\item Multiple Choice Question.
-\begin{enumerate}
-\item Choice
-\item Choice
-\item Choice \(\checkmark\)
-\end{enumerate}
-
-\item Multiple Choice Question.
-\begin{enumerate}
-\item Choice
-\item Choice
-\item Choice \(\checkmark\)
-\end{enumerate}
-\end{enumerate}
-\end{document}
