+++ Courses/Org-Coursepack-Quickstart-Guide/2018 Fall/Syllabus/Syllabus (Section 1).tex
@@ -1,123 +1,0 @@
-% Intended LaTeX compiler: pdflatex
-\documentclass[article,letterpaper,times,10pt,listings-bw,microtype]{scrartcl}
-\usepackage[utf8]{inputenc}
-\usepackage[T1]{fontenc}
-\usepackage{graphicx}
-\usepackage{grffile}
-\usepackage{longtable}
-\usepackage{wrapfig}
-\usepackage{rotating}
-\usepackage[normalem]{ulem}
-\usepackage{amsmath}
-\usepackage{textcomp}
-\usepackage{amssymb}
-\usepackage{capt-of}
-\usepackage{hyperref}
-\usepackage[onehalfspacing]{setspace} \usepackage[T1]{fontenc} \usepackage{mathpazo} \usepackage[scaled=0.85]{helvet} \usepackage{courier} \usepackage{geometry} \geometry{left=1in,right=1in,top=1in,bottom=1in} \usepackage[framemethod=default]{mdframed}
-\author{Your Name Your Name}
-\date{Fall 2018}
-\title{COURSE 0000-01 Syllabus}
-\hypersetup{
- pdfauthor={Your Name Your Name},
- pdftitle={COURSE 0000-01 Syllabus},
- pdfkeywords={},
- pdfsubject={Org-Coursepack quickstart guide School specific teaching materials},
- pdfcreator={Emacs 26.1 (Org mode 9.1.14)}, 
- pdflang={English}}
-\begin{document}
-
-\definecolor{SchoolColor}{RGB}{70,130,180}
-
-\begin{center}
-{\color{SchoolColor}{\Large
-\textbf{School Name}
-}}
-\end{center}
-
-\begin{center}
-{\color{SchoolColor}{
-\textbf{COURSE 0000-01}
-
-\textbf{Org-Coursepack Quickstart guide}
-
-\textbf{Fall 2018}
-}}
-\end{center}
-
-\vspace{5 mm}
-
-\begin{center}
-\begin{tabular}{llll}
-\textbf{Instructor:} & Your Name & \textbf{Office Phone:} & (000) 000-0000\\
-\textbf{Office:} & BLDG 100 & \textbf{E-mail:} & YourEmail\\
-\textbf{Office Hours:} & Tue 3:30-4:30pm & \textbf{Course Site:} & \textbf{\href{https://github.com}{github.com}}\\
-\textbf{Class Meeting Day \& Time:} & Tue/Thurs, 9:30a-10:45 & \textbf{Class Location:} & BLDG 100\\
-\end{tabular}
-\end{center}
-\section*{Course Description}
-\label{sec:orgba742b8}
-This is a course intended as the quickstarter guide for \texttt{Org-Coursepack}. We
-aim to equip students with the basic usage of \texttt{Org-Coursepack}.
-\section*{Course Prerequisites}
-\label{sec:org4c7a1f4}
-\begin{itemize}
-\item Course Prerequisites here.
-\end{itemize}
-\section*{Student Learning Objectives}
-\label{sec:org19b521a}
-As the result of this course, students should be able to:
-
-\begin{itemize}
-\item Student Learning Objectives here.
-\end{itemize}
-\section*{Course Material}
-\label{sec:org1858e2f}
-\begin{itemize}
-\item Course Materials here.
-\end{itemize}
-\section*{Grading}
-\label{sec:org2195b5d}
-Grading information here.
-\subsection*{Grade Cutoff Percentage}
-\label{sec:org9788cad}
-The following table provides the preliminary guidelines for assigning your
-letter grade:
-
-\begin{center}
[…]
