### 2018-12-25 #34 @joonro
Dear @pschloss and @gregcaporaso,

We have made significant updates to make Org-Coursepack more approachable to a wider audience, even those who are new to OrgMode or Emacs. Please see the updated [repo](https://github.com/joonro/Org-Coursepack). Specifically, based on your feedback, we focused on the following two things.

First was creating the [Quickstart guide](https://joonro.github.io/Org-Coursepack/Lectures/01%20Course%20Introduction.html#quickstart-guide). This guide provides step-by-step instructions with screenshots of examples so that users can jumpstart a new course construction.

Second, we substantially improved the installation instructions based on your comments. As a part of this update, we added an example init file for Emacs, which (a) includes basic user settings and (b) automatically installs necessary packages required to use the `Org-Coursepack`. We apologize for not including these instructions previously.

We realized from your valuable comments the importance of making the Org-Coursepack more approachable to various levels of users, and thus, have made a concerted effort to achieve this. Please let us know if you have any other comments or suggestions. Thank you for your time, and happy holidays!
