### 2018-12-25 joonro/Org-Coursepack#3 @joonro
Dear @gregcaporaso,

As I replied in the [review page](https://github.com/openjournals/jose-reviews/issues/34#issuecomment-449792558), we added quickstart guide as well as more comprehensive installation instructions to the repo. Importantly, we also added an [example init file](https://github.com/joonro/Org-Coursepack/blob/master/Assets/.emacs.d/init.el), which will automatically take care of installation of necessary Emacs packages along with containing necessary settings.

We believe now users can easily install and start using Org-Coursepack. Please let me know if you need anything else regarding this issue, and thank you again for your feedback.
