### 2019-01-08 #34 @pschloss
Yes

Sent frim my iPhone, expext more typos then nirmal

> On Jan 8, 2019, at 15:26, Lorena A. Barba <notifications@github.com> wrote:
>
> @pschloss You have a few unchecked items in the checklist. Are you checking these off with your recommendation now?
>
