### 2019-01-11 joonro/Org-Coursepack#1 @gregcaporaso
I would recommend some sort of tests that confirm that the products can be created from the source content, but since I'm not an Org user, I'm not sure that this is possible or something that makes sense to add.
