### 2019-01-11 #34 @gregcaporaso
The new quickstart guide looks great - thanks @joonro!

There are two items in my list that I can't check off yet:

> Version: Does the release version given match the GitHub release (v1.0.1)?

The version listed on this page is 1.0.1, but the most recent version of the project is 1.1.0. I suspect this just needs to be updated on this page.

> Tests: Are there automated tests or manual steps described so that the function of the software can be verified?

There are not automated tests - this came up in https://github.com/joonro/Org-Coursepack/issues/1. The quickstart guide could be viewed as tests that can be run manually, though that may be too time consuming to reasonably be run every time a PR is issued or merged. Do other Org projects have automated tests associated with them? Matching the testing approach for other Org projects is probably what should be considered sufficient for this project.
