### 2019-01-11 #34 @joonro
Thank you @gregcaporaso. Regarding the version number difference, I don't think I can update the issue. I was wondering if that is something you can modify so it reflects the number on the repo?  @labarba, Please let me know if I need to do something to update it.

We will look into incorporating some automated test and report back soon!
