### 2019-01-11 #34 @labarba
Our lovely bot `whedon` can change the version 😁
