### 2019-01-12 a3b38ac
Add .travis.yml, a modified version of https://github.com/rolandwalker/emacs-travis/blob/master/.travis.yml
