### 2019-01-12 aed2bd3
Try using a docker image for Emacs
