+++ .travis.yml
@@ -1,132 +1,11 @@
-###
-### Notes
-###
-### The travis web interface may choke silently and fail to
-### update when there are issues with the .travis.yml file.
-###
-### The "travis-lint" command-line tool does not catch all
-### errors which may lead to silent failure.
-###
-### Shell-style comments in this file must have "#" as the
-### *first* character of the line.
-###
+language: emacs-lisp
 
-###
-### language
-###
-
-# travis-lint no longer permits this value
-# language: emacs-lisp
-
-###
-### defining the build matrix
-###
-### ===>                                                       <===
-### ===> each variation in env/matrix will be built and tested <===
-### ===>                                                       <===
-###
-### variables under env/global are available to the build process
-### but don't cause the creation of a separate variation
-###
-
-env:
-  matrix:
-    - EMACS=emacs25
-    - EMACS=emacs26
-    - EMACS=emacs-snapshot
-  global:
-    - SOME_TOKEN=some_value
-
-###
-### allowing build failures
-###
-
-matrix:
-  allow_failures:
-    - env: EMACS=emacs-snapshot
-
-###
-### limit build attempts to defined branches
-###
-### notes
-###
-### This controls which branches are built.
-###
-### You can also control which branches affect the web badge, by
-### appending "?branch=master,staging,production" to the end of the
-### image URL (replacing "master,staging,production" with a
-### comma-separated list of branches to be reflected in the badge).
-###
-#
-#  branches:
-#    only:
-#      - master
-#
-
-###
-### runtime initialization
-###
-### notes
-###
-### emacs-snapshot (trunk) is obtained from the Ubuntu Emacs Lisp PPA:
-### https://launchpad.net/~ubuntu-elisp/+archive/ppa
-### For the emacs-snapshot build, bleeding-edge versions
-### of all test dependencies are also used.
-###
+services:
+  - docker
 
 before_install:
-  - git submodule --quiet update --init --recursive
-
-install:
-  - if [ "$EMACS" = 'emacs25' ]; then
-        sudo add-apt-repository -y ppa:kelleyk/emacs &&
-        sudo apt-get -qq update &&
-        sudo apt-get -qq -f install &&
-        sudo apt-get -qq install emacs25 emacs25-el;
-    fi
-  - if [ "$EMACS" = 'emacs26' ]; then
-        sudo add-apt-repository -y ppa:kelleyk/emacs &&
-        sudo apt-get -qq update &&
-        sudo apt-get -qq -f install &&
-        sudo apt-get -qq install emacs26 emacs26-el;
-    fi
-  - if [ "$EMACS" = 'emacs-snapshot' ]; then
-        sudo add-apt-repository -y ppa:ubuntu-elisp/ppa &&
-        sudo apt-get -qq update &&
-        sudo apt-get -qq -f install &&
-        sudo apt-get -qq install emacs-snapshot &&
-        sudo apt-get -qq install emacs-snapshot-el;
-    fi
-
-before_script:
-  - if [ "$EMACS" = 'emacs-snapshot' ]; then
-        make downloads-latest;
-    else
-        make downloads;
-    fi
-
-###
-### the actual build/test command
-###
-### Use "make test-batch" to test without byte-compiling.
-### The default command avoids byte-compiling on Emacs 22.
-###
+- docker docker pull silex/emacs
+- docker run -it --rm silex/emacs
 
 script:
-  $EMACS --batch -l tests.el 
-
-###
-### settings
-###
-
-notifications:
-  email: false
-
-#
-# Emacs
-#
-# Local Variables:
-# indent-tabs-mode: nil
-# coding: utf-8
-# End:
-#
+- emacs --batch -l tests.el
