+++ .travis.yml
@@ -4,7 +4,7 @@ services:
   - docker
 
 before_install:
-- docker docker pull silex/emacs
+- docker pull silex/emacs
 - docker run -it --rm silex/emacs
 
 script:
