### 2019-01-12 ce564a1
Run emacs in the batch mode
