### 2019-01-12 15d5c04
run emacs directly
