### 2019-01-13 f995524
Docker setup update
