+++ .travis.yml
@@ -5,6 +5,10 @@ services:
 
 before_install:
 - docker pull silex/emacs
+- docker run -dt --name emacs silex/emacs /bin/bash
+- docker cp ./Org-Coursepack emacs:/home/Org-Coursepack
+- docker exec -i -t emacs /bin/bash
+- cd /home/Org-Coursepack
 
 script:
-- docker run -it --rm silex/emacs emacs --batch -l tests.el
+- emacs --batch -l tests.el
