+++ .travis.yml
@@ -6,7 +6,7 @@ services:
 before_install:
 - docker pull silex/emacs
 - docker run -dt --name emacs silex/emacs /bin/bash
-- docker cp ./Org-Coursepack emacs:/home/Org-Coursepack
+- docker cp ../Org-Coursepack emacs:/home/Org-Coursepack
 - docker exec -i -t emacs /bin/bash
 - cd /home/Org-Coursepack
 
