+++ .travis.yml
@@ -7,8 +7,7 @@ before_install:
 - docker pull silex/emacs
 - docker run -dt --name emacs silex/emacs /bin/bash
 - docker cp ../Org-Coursepack emacs:/home/Org-Coursepack
-- docker exec -i -t emacs /bin/bash
-- cd /home/Org-Coursepack
+- docker exec -i -t emacs /bin/sh -c "cd /home/Org-Coursepack && emacs --batch -l tests.el"
 
 script:
 - emacs --batch -l tests.el
