### 2019-01-13 3b6f564
Try vanila emacs
