+++ .travis.yml
@@ -1,9 +1,11 @@
+sudo: required
+
 language: emacs-lisp
 
 before_install:
 - sudo add-apt-repository ppa:kelleyk/emacs -y
 - sudo apt-get update -q
-- sudo apt install -y emacs26
+- sudo apt install -y emacs
 
 script:
 - emacs --batch -l tests.el
