+++ .travis.yml
@@ -5,7 +5,7 @@ language: emacs-lisp
 before_install:
 - sudo add-apt-repository ppa:kelleyk/emacs -y
 - sudo apt-get update -q
-- sudo apt install -y emacs texlive-base
+- sudo apt install -y emacs texlive
 
 script:
 - emacs --batch -l tests.el
