### 2019-01-13 #34 @joonro
Dear @gregcaporaso,

We just finished adding automated tests to the repo using Travis CI as you requested. They include exporting of all course contents (syllabus, lecture, exam, and assignment) from the semester template Org file. In addition, the tests load the example init file before exporting contents and thus the validity of the init file (Emacs package installation and other settings) is automatically checked as well.

Please let me know if you need anything else, and thank you very much.
