### 2019-01-13 #34 @joonro
I just updated the repo with improved testing - now it also includes testing of `org-coursepack-update-lecture-metadata` script as well as updating of class schedule table.
