+++ Courses/Org-Coursepack/2018 Fall/Lectures/08 Semester Org Files 3 4 Lectures.rst
@@ -3,8 +3,11 @@ Class 8: Semester Org Files (3/4): Lectures
 ===========================================
 
 
- ORG 0000  Joon H. Ro & Jae-Eun Namkoong  
- Class 8  [2018-09-20 Thu] 
+ORG 0000   
+Joon H. Ro & Jae-Eun Namkoong 
+
+Class 8  
+[2018-09-20 Thu]
 
 1 The ``Lectures`` Top-level Tree
 ---------------------------------
@@ -18,8 +21,7 @@ Class 8: Semester Org Files (3/4): Lectures
       * Lectures
       :PROPERTIES:
 
-      #+NAME: Update Classes
-      #+BEGIN_SRC emacs-lisp :results none
+      #+CALL: org-coursepack-update-lecture-metadata()
 
       ** Tasks [0/1]                                           :noexport:skipcount:
       ** Classes and Assignments Dates                                  :skipcount:
@@ -50,67 +52,68 @@ Class 8: Semester Org Files (3/4): Lectures
 
       :PROPERTIES:
       :COLUMNS: %Date %Class %ITEM
-      :ID:  79d5e887-4637-43e7-8e8a-b83fa83ee56e
+      :ID:  Lectures
       ...
       :END:
 
-1.2 ``Update lecture metadata`` Source Code Block
-~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+1.2 ``org-coursepack-update-lecture-metadata`` Source Code Block
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
-- The ``Lectures`` top-level tree has a source code block named ``Update Lectures``. When executed with ``C-c C-c``, it is designed to go through
-  each lecture subtree and perform the following actions:
+- The ``Lectures`` top-level tree has a call statement (``#+CALL: org-coursepack-update-lecture-metadata()``) that can be used to execute the
+  source code block named ``org-coursepack-update-lecture-metadata``, which is
+  contained in ``/Assets/Scripts.org``. When executed with ``C-c C-c``, it is
+  designed to go through each lecture subtree and perform the following
+  actions:
 
-  1. Update ``:PROPERTIES:`` for the lecture, such as the class number
-     (``:CLASS:``), class date (``:DATE:``), and file name of the export
-     (``:EXPORT_FILE_NAME:``).
+  1. Update metadata for the lecture, such as the class number (``:CLASS:``),
+     class date (``:DATE:``), and file name of the export
+     (``:EXPORT_FILE_NAME:``), which are defined in the ``:PROPERTIES:``.
 
      ``:CLASS:``
-         The lectures are assumed to be in the order of the
-         lecture schedule (e.g., first lecture on Class
-         1). Note that any subtree with a ``:skipcount:`` tag
-         will be ignored, which is useful for
-         non-lecture subtrees (e.g., assignment due dates
-         and holidays).
+         The lectures are assumed to be in the order of the lecture
+         schedule (e.g., first lecture on Class 1). Note that any
+         subtree with a ``:skipcount:`` tag will be ignored, which is
+         useful for non-lecture subtrees (e.g., assignment due
+         dates and holidays).
 
      ``:DATE:``
-         It will get the date of each class from the ``DATE_CLASS_XX`` file-level
-         properties, which are defined in the ``Lectures and Assignments Dates`` subtree.
+         It will get the date of each class from the ``DATE_CLASS_XX``
+         file-level properties, which are defined in the ``Lectures and Assignments Dates`` subtree.
 
      ``:EXPORT_FILE_NAME:``
[…]

[…]
