### 2019-01-14 #34 @gregcaporaso
All changes look good, I recommend publication.
