### 2019-01-14 #34 @labarba
@joonro — I do have one editorial request: that you add a section heading *Statement of Need*, as described in the [Author Guidelines](http://jose.theoj.org/about#author_guidelines). You can put under that heading the paragraphs in your paper that address the need for `Org-Coursepack`.
