+++ paper.md
@@ -24,64 +24,67 @@ bibliography: paper.bib
 # Summary
 
 The Org-Coursepack provides a template for developing and managing teaching
-materials using [Org mode](https://orgmode.org), a major mode in [GNU Emacs](https://www.gnu.org/software/emacs/). The template is designed
-to be self-explanatory, as the documentation for it consists of an example
-course developed with this template.
+materials using [Org mode](https://orgmode.org), a major mode in [GNU
+Emacs](https://www.gnu.org/software/emacs/). The template is designed to be
+self-explanatory since the documentation for it consists of a sample course
+developed with this template.
+
 # Statement of Need
 
 Org-Coursepack started as an attempt to solve challenges authors experienced
-in creating and managing teaching materials. The key benefits of using the
+when creating and managing teaching materials. The key benefits of using the
 template as an instructor are summarized below.
 
 First, sharing materials across courses becomes more effective and efficient
-with the template than with traditional means of creating lecture slides (e.g.,
-PowerPoints or Beamer). This is because the template heavily relies on the
-file inclusion functionality of Org mode and is designed to be as modular as
+with the template than with traditional means of creating lecture slides
+(e.g., PowerPoints or Beamer) because the template relies heavily on the file
+inclusion functionality of Org mode; thus, it is designed to be as modular as
 possible. As a result, a course topic can be used across multiple sections of
-the same course or across multiple courses—this means that making improvements
-on one lecture file will be automatically propagate across all sections and
-courses covering the topic, reducing redundancy and potential for
-inconsistency. Thus, Org-Coursepack would be especially useful when an
-instructor has multiple courses which share some common course materials.
+the same course or across multiple courses, which means improvements made to
+one lecture file will automatically propagate across all sections and courses
+covering the topic, reducing redundancy and inconsistency. Thus,
+Org-Coursepack would be especially useful for an instructor with multiple
+courses that share common course materials.
 
-Second, the flexible export system and output-specific export options of Org
+Second, the flexible export system and output-specific export options in Org
 mode allow instructors to generate outputs in multiple formats. Any
-course-related material, whether a syllabus, exam review, or lecture, can be
-easily exported in the form of slides (e.g., via Reveal.js or Beamer backends)
-or handouts (e.g., via the LaTeX backend) while maintaining consistent content
-regardless of format. This allows instructors to create properly formatted,
-document-like handouts for their lectures, which have better presentation and
-legibility compared to scaled-down slides. In addition, instructors can choose
-to selectively present certain components (e.g., images, charts, or notes)
-depending on output format. These features eliminate the need for instructors
-to maintain separate files for different output formats.
+course-related material, such as syllabi, exam reviews, or lectures, can be
+easily exported in the form of slides (e.g., via reveal.js or Beamer backends)
+or handouts (e.g., via LaTeX or reStructuredText backends) while maintaining
+consistent content regardless of format. This allows instructors to create
+properly formatted, document-like handouts for their lectures, offering better
+presentation and legibility compared to scaled-down slides. In addition,
+instructors can choose to selectively present certain components (e.g.,
+images, charts, or notes) depending on the output format. These features
+eliminate the need for instructors to maintain separate files for different
+output formats.
 
-Third, the template contains a) utility functions written in Emacs Lisp, b)
-shortcuts to Org mode functions, and c) basic class structures, all of which
+Third, the template contains utility functions written in Emacs Lisp,
+shortcuts to Org mode functions, and basic class structures, all of which
 contribute to automating a variety of instruction-related tasks. The features
 include, but are not limited to, automatic class numbering and automatic
-creation of key content, such as course schedule for syllabi, summary of class
-materials, and exam keys.
+creation of key content, such as course schedules for syllabi, summaries of
+class materials, and exam keys.
 
[…]
