### 2019-01-14 #34 @joonro
Dear @labarba, I'm so sorry, I just noticed that the second author's first name is missing a hyphen (should be Jae-Eun Namkoong instead of Jae Eun Namkoong). Would it be possible to change this?
