### 2019-01-14 #34 @labarba
Did you fix the paper source?

@arfon — Can you help me? I'm not sure what to do here: the author wants a hot fix on the author name. Can we update the PDF with `@whedon generate pdf`? Do we have to do something else with the Crossref deposit if the paper is fixed after publication?
