### 2019-01-15 #34 @arfon
> @arfon — Can you help me? I'm not sure what to do here: the author wants a hot fix on the author name. Can we update the PDF with @whedon generate pdf? Do we have to do something else with the Crossref deposit if the paper is fixed after publication?

Unfortunately Whedon doesn't currently support updates like this. If you still have Whedon installed locally then you'll need to do the download and compile steps locally. I can then update the repos with the latest PDF and update the Crossref metadata.
