### 2019-01-15 #34 @labarba
Thanks, @arfon — I used `whedon` locally to re-download the submission repository to `whedon/tmp`, I re-compiled with `bundle exec whedon compile 34`, copied the files over to the local  `jose-papers` repo in the correct folder, then pushed to the remote (both the PDF and XML files).

Should I now run `bundle exec whedon deposit 34` locally, and that will do it?
