### 2019-01-15 #34 @labarba
BTW, I got this warning after the compilation:

```
[WARNING] Missing character: There is no ― in font [lmroman10-regular]:mapping=tex-text;!
tmp/34/10.21105.jose.00034.pdf
```

... but the PDF looks OK to me.
