### 2019-01-15 #34 @arfon
> Should I now run `bundle exec whedon deposit 34`

Yes, that should do it.
